add foreign keys for project tables

+++ sql_project/V3_4__car_showroom_create.sql
@@ -1,4 +1,4 @@
 CREATE TABLE `car_showroom` (
-  `id` INT NOT NULL AUTO_INCREMENT,
+  `id` SMALLINT NOT NULL AUTO_INCREMENT,
   `address_id` SMALLINT NOT NULL,
   PRIMARY KEY (`id`));
