+++ sql_project/V3_6__privod_create.sql
@@ -1,4 +1,4 @@
 CREATE TABLE `privod` (
-  `id` INT NOT NULL AUTO_INCREMENT,
+  `id` TINYINT NOT NULL AUTO_INCREMENT,
   `privod_type` VARCHAR(50) NOT NULL,
   PRIMARY KEY (`id`));
